Change the re-lock behavior to allow shorter extensions.
The new lock end date after the re-lock attempt must exceed previous end date, so the lock length can not be shortened. But the new duration can be shorter than the previous one. But since the rewards are based on duration length, we want to re-calculate duration before storing it so it reflects the extended lock time from now on.