Use the byte3 array directly for the versioning
The Version contract defined the version as the string "305" and the function returns a bytes3 value. The string "305" was encoded and the returned as 0x333035. This limited patch increments to the versioning. For example, we would not be able to fit the string "3010" (for version 3.0.10) into bytes3.

The update changes the version encoding to be simply 3 digits, base 256 number. This allows major, minor, and patch numbers up to 255.

The update was suggested by OpenZeppelin reviewer; no issue reference was assigned.